--- .dockerignore
@@ -1,20 +1,0 @@
-**/.dockerignore
-**/.env
-**/.git
-**/.gitignore
-**/.vs
-**/.vscode
-**/*.*proj.user
-**/azds.yaml
-**/charts
-**/bin
-**/obj
-**/Dockerfile
-**/Dockerfile.develop
-**/docker-compose.yml
-**/docker-compose.*.yml
-**/*.dbmdl
-**/*.jfm
-**/secrets.dev.yaml
-**/values.dev.yaml
-**/.toolstarget
