+++ .gitignore
@@ -8,6 +8,7 @@
 *.user
 *.userosscache
 *.sln.docstates
+*.bat
 
 # User-specific files (MonoDevelop/Xamarin Studio)
 *.userprefs
