--- ConfigMapFileProviderSample.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29215.179
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConfigMapFileProviderSample", "src\ConfigMapFileProviderSample\ConfigMapFileProviderSample.csproj", "{1343A1BF-33E0-49D5-AADB-A323867697AE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1343A1BF-33E0-49D5-AADB-A323867697AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1343A1BF-33E0-49D5-AADB-A323867697AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1343A1BF-33E0-49D5-AADB-A323867697AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1343A1BF-33E0-49D5-AADB-A323867697AE}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {AB82D515-15E8-4AAE-A7B7-0CDC78DB7234}
-	EndGlobalSection
-EndGlobal
