+++ src/ConfigMapFileProvider/ConfigMapFileProvider.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <Description>
+    Dynamic reload config from file for Kubernetes with Config Map. 
+    Usage:
+      configBuilder.SetBasePath(Path.Join(AppContext.BaseDirectory, "config"))
+                  .AddJsonFile(ConfigMapFileProvider.FromRelativePath("config"), "appsettings.json", true, true);
+    </Description>
+    <version>1.0.0</version>
+    <authors>Jaine.ch</authors>
+    <owners>Jaine.ch</owners>
+    <releaseNotes>First version</releaseNotes>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.FileProviders.Abstractions" Version="3.1.9" />
+    <PackageReference Include="Microsoft.Extensions.FileProviders.Physical" Version="3.1.9" />
+    <PackageReference Include="Microsoft.Extensions.Primitives" Version="3.1.9" />
+  </ItemGroup>
+
+</Project>
