--- src/ConfigMapFileProviderSample/configmap.yaml
@@ -1,15 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: demo-config
-data:
-  appsettings.json: |-
-    {
-      "Logging": {
-        "LogLevel": {
-          "Default": "Error",
-          "System": "Error",
-          "Microsoft": "Error"
-        }
-      }
-    }
