--- src/ConfigMapFileProviderSample/deployment.yaml
@@ -1,29 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: demo-deployment
-  labels:
-    app: config-demo-app
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: config-demo-app
-  template:
-    metadata:
-      labels:
-        app: config-demo-app
-    spec:
-      containers:
-      - name: configmapfileprovidersample
-        imagePullPolicy: Always
-        image: fbeltrao/configmapfileprovidersample:1.0
-        ports:
-        - containerPort: 80
-        volumeMounts:
-        - name: config-volume
-          mountPath: /app/config
-      volumes:
-      - name: config-volume
-        configMap:
-          name: demo-config
